+++ releasenotes.md
@@ -1,3 +1,5 @@
 Whats New in v12.6
 --------------------------
+- Added meta data (formulas, ingredients,...) count for users in the admin dashboard
+
 - This release may include more changes, for full details please refer to the CHANGELOG